+++ src/client/java/tempeststudios/quickstacknearby/QuickStackNearbyClient.java
@@ -5,6 +5,7 @@ import net.fabricmc.api.ClientModInitializer;
 public final class QuickStackNearbyClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
+        QuickStackHotkey.register();
         QuickStackNearbySmokeTest.registerIfEnabled();
     }
 }